chacha20_poly1305: Split integrated/non-integrated impls.
Split up the implementations so that future refactorings to the
CPU-feature-based dispatching logic will be easier.

The input length checking is done in every function since it
potentially affects the correctness and memory safety of each
assembly language function, and we want to encapsulte the call to
each `unsafe` assembly language function with a safe wrapper.

```
git difftool HEAD^1:src/aead/chacha20_poly1305.rs \
                    src/aead/chacha20_poly1305/mod.rs
git difftool HEAD^1:src/aead/chacha20_poly1305.rs \
                    src/aead/chacha20_poly1305/integrated.rs
```
